+++ lib/interfaces/TempSensor.h
@@ -9,8 +9,8 @@ class TempSensor {
         /**
          * @brief Read the temperature.
          * 
-         * @returns A double of the value of the temperature in degrees Celsius.
+         * @returns A float of the value of the temperature in degrees Celsius.
          */
-        virtual double read_temp() = 0;
+        virtual float read_temp() = 0;
         virtual ~TempSensor() = default;
 };
